(SERVER-1787) Puppet 3 agents must use foss install type
If we're running Puppet 3 compatibility tests the agent install type
doesn't matter, as we _must_ use the 'foss' install type. This commit
manually reconfigures the Puppet agent hosts to use the 'foss' install
type and ignore the '--aio' command line flag and config option as that
option can only meaningfully apply to the master node, and never the
agents under test.