(SERVER-1786) Install Puppet GPG key for SLES 12 Puppet 4 compat agent
Previously, when the Puppet 4 compatibility suite tried to install the
SLES 12 agent from the PC1 repo, it failed due to the GPG public key
associated with the package not being present on the host.  This commit
adds a step which installs the public key prior to the package install.
This is done for SLES 12 only because it isn't necessary for other
package managers.  yum and apt installs pull down the public key
automatically as part of the package install.