(SERVER-911) Remove sync step for PE in reload crl test
Previously, the `reload_updated_crl` acceptance test included a step for
PE which copied the file at the cacrl location over to the file at the
hostcrl location after a certificate was revoked.  The work done for
PE-19484 allowed for the hostcrl file to be updated automatically after
service restart for PE, as had already previously been the case for OSS
Puppet Server.  This commit removes the synchronization step from the
test since it is no longer necessary for PE.