(SERVER-715) environment registry isExpired returns nil
The EnvironmentRegistry.isExpired method reified in the environment-registry fn will
try to return nil if an environment is unregistered. The java interface requires it
returns a boolean, and so an exception is thrown.

This commit changes the implementation of isExpired to return true if the environment
is unregistered, since returning false might indicate the cache entry is still valid,
when the registry doesn't actually know anything about the entry