(SERVER-1827) Change default cert-status content-type to json
This commit changes the default content-type which is returned when no
'accept' header is given for a certificate-status request from
'text/pson' to 'application/json'.  This is being done for consistency
with defaulting to 'application/json' in place of 'text/pson' elsewhere
for server endpoints in Puppet 5.  Support for roundtripping 'accept:
pson' or 'accept: text/pson' requests in responses as 'content-type:
pson' or 'content-type: text/pson', respectively, has been preserved.