(MAINT) Replace archive_files with autoflush for settings test
Previously, the Clojure unit tests contained a test for the value of the
Puppet archive_files setting.  Per changes in PUP-7260 which just landed
in Ruby Puppet, this setting no longer exists.  This commit changes the
test to instead validate the value of the autoflush setting.  The
specific setting being tested isn't important.  The test just proves
that a known expected default value for a setting can be retrieved from
Ruby Puppet.