(MAINT) Added logic to code_scripts.rb test to make it re-runnable
This commit includes some modifications to the code_scripts.rb
acceptance test which allow it to be run multiple times, as may be
needed for easier troubleshooting.  The commit uses temp directories
instead of hard-coded directories in a couple of spots and ensures that
proper directory ownership is set for some directories.