(SERVER-1825) Add puppet rubygems mirror for bundler install
This commit honors the GEM_SOURCE environment variable as an extra gem source
for the bundler gem install done for JRuby spec testing.  If the
variable is not set, the internal puppet rubygems mirror is added as an
extra default gem source.  This should help reduce test run failures due
to intermittent issues seen when accessing the rubygems.org server.