(maint) Switch java version on Debian 8 upgrade tests
When testing the upgrade path on Debian 8, Puppetserver 2.7.x pulls
in openjdk-7-jre-headless as a dependency. Upgrading Puppetserver to 5.x
pulls in openjdk-8-jre-headless but this Java version isn't necessarily
used. This commit adds a call to update-alternatives on Debian 8 to
ensure that we have the right jvm installed and active for testing.