(SERVER-1866) Copy the cacrl to the hostcrl file on change at runtime
This commit uses the tk-filesystem-watcher service to listen for changes
to the cacrl file.  When a change is detected, the cacrl file is copied
over to the hostcrl file.  This allows a watcher that the newer
tk-jetty9 service has established on changes to the hostcrl file to
trigger a reload of the CRL being used by the webserver, with no restart
to the webserver service being required.

The tk-jetty9 bump in this PR for the crl watcher changes also includes
a fix for previously failing gzip encoding for server responses.
To account for this fix, a couple of updates to the
environment-classes-int-test tests were added to account for the --gzip
text appended to some response etags.