(SERVER-1894) install `apt-transport-https` on Debian for PDB
We install PDB and smoke test it in
`acceptance/suites/tests/00_smoke/puppetdb_integration.rb`. As of
2017-07-17 the puppetlabs-puppetdb module will attempt to install
postgresql using a repo with in the https url. apt-get cannot use
https by default, the `apt-transport-https` package installs the addon
needed to use apt-get with https.

This patch installs `apt-transport-https` on debian variants.