(SERVER-1866) Remove timestamp check and reduce sleep
This commit addresses two pieces of feedback from the PR. First, the
timestamp check had the potential to cause this section of code to
indefinitely loop, given the right outside process. Second the sleep
seemed too long for write operations on the cacrl that might hold a lock
on the file.
The first issue is handled by simply removing the modification check,
trusting that if the modification happens after the loop has ended, a
new event will be triggered that will cause the copy to be retried. The
second issue is addressed by reducing the sleep to 100 milliseconds and
to increase the number of attempts to 25.