(maint) delete puppet version spec test
Previously we checked the version of Puppet from PuppetServer's Ruby
interface. This test was checking that we could reach values that Puppet
sets in Ruby via JRuby running through Puppet Server's JVM layer.

Testing the version makes any kind of automated version bumping
problematic. There is a similar test that checks run_mode so the risk
in a bug not being caught by this test is small, and the return on
investment in automating some of our workflow is large.

This patch deletes the spec test.