(SERVER-1755) Do not deserialize msgpack payloads to UTF-8 string
Msgpack data, like other binary data, cannot be deserialized to a UTF-8
string. In general, we try to decode all payloads this way when we
receive requests. This commit adds an exception for payloads from
msgpack, sending them as is to Ruby Puppet.

This also updates the unit test concerning binary content types.