(SERVER-1920) Use newer grafandash module for graphite integration
Previously we used a rarely updated in three years modules from an
employee that no longer works here.

That module is broken and the target OS for it is out of date (CentOS
6).

This patch updates the tests to use a new module within the puppetlabs
namespace and target CentOS 7.