(SERVER-1315) Extract and verify ca cert from bundle
In order to support certificate bundles the `cacert` file has changed
from just holding a single certificate to potentially holding a
certificate chain. When we need a copy of the CA certificate we need to
extract the first cert from the bundle and verify that it matches the
signing key. This commit relies on the `pem->ca-cert` function to
extract and verify the cert whenever we need to touch the CA
certificate.