(SERVER-249) Update cli classpath to include jar directory
In order to support adding arbitrary JARs to puppetserver (e.g. for use
in custom facts or to replace native gem extensions), we have added a
directory to puppetserver's data directory, where users can place JARs.
This commit updates the classpath specified in the script that sets
environment variables prior to running puppetserver subcommands to
include this new location.