(maint) Remove bindaddress config tests
Some clojure tests were inspecting setting information coming from
Puppet and Ruby to verify that the information could be fetched, but
this failed when Puppet itself changed a default value. This commit
changes the tested values to another string type value to check the same
code paths so we don't have to have tests depending on what version
of Puppet is in use.