(TK-149, SERVER-1866) Update acceptance test to enable CRL reloading
Previously, we were testing whether or not CRL reloading worked in our
default configuration.

With recent work this only happens if explicitly enabled, recommended
only on Java 8.

This updates the test to enable the CRL reloading before attempting to
test for it, and only if the server is running Java 8.