(TK-149) remove beaker test duplicated by clj test
We have clojure integration tests that exercise the CRL reload functionality
quite throughly. After this acceptance test landed, we realized we could not
support this feature on java 1.7, resulting in additional complexity around
testing only on Java 8. Given this test is largely duplicate and the additional
complexity needed to maintain it, this commit opts to remove it instead and
leverage the clojure integration tests for validation.

Signed-off-by: Moses Mendoza <moses@puppet.com>