(SERVER-1315) Sync up bundle certificates
The Puppetserver CA implementation does not verify that a certificate
being revoked was actually signed by the CA, which caused the tests to
try to revoke a certificate that could not be verify as revoked in the
tests. This commit rebuilds the test fixtures to ensure that the
test certificate being revoked actually can be verified as revoked.