(SERVER-1925) Integrate endpoint with JRuby
Add integration smoke test for endpoint. Details have mostly been
tested in the unit test in previous commit, but this test makes sure the
system works end-to-end.

Turned out we needed `sort-nested-info-maps` to be able to handle the
output from the JRubyService -- the returned maps are RubyHash objects,
and values can't be accessed in normal Clojure ways. So call
`sort-nested-info-maps` before passing task information to the functions
that format it, so that all-tasks-response! can accept both Clojure data
structures and JRuby ones.