(SERVER-1925) Preliminary docs
Write some preliminary docs for the /puppet/v3/tasks endpoint. These
aren't perfect yet; we're going to wait until the task detail endpoint
docs are written to get edits from the Docs team, so the goal for this
commit is just to have docs that explain the endpoint in an
understandable enough way.

I also expect when other tasks documentation gets written we'll want to
link to some of that in the final docs.