(SERVER-1933) Copy cacrl to hostcrl when cacrl file changes
The latest version of trapperkeeper-webserver-jetty9 enabled hot
reloading of the CRL. Whenever this reload happens, we want to copy
those changes to our hostcrl. This commit uses the filesystem-watcher to
watch for changes to the cacrl file and trigger an update to our hostcrl
when they occur.

This functionality can be disabled by not loading the filesystem-watcher
service.

This feature has been ported from the 2.x branch.