(SERVER-1933) Make automatic CRL reload mandatory
While in 2.x it was necessary to support disabling automatic CRL reload
due to incompatability with Java 7, for Puppet Server 5 we want this
feature to be mandatory, since we do not support Java 7. This commit
removes the logic that made automatic reloading optional, as well as the
accompanying tests and docs.