(SERVER-1973) Error on unparseable metadata
Alter the task detail endpoint to return an error if the JSON metadata
for a task is unparseable.

This change is necessary for security reasons: if a task's metadata
restricts the allowed parameters to a task, and there's a typo preventing
that metadata from being parsed, we don't want people to be able to run
the task as if there are no restrictions in place (as would be the case
with an empty metadata object).

Getting the tests to passed required setting the environment_timeout to
zero, as it doesn't create all tasks at once (and an environment_timeout
of unlimited will cache what modules and tasks are found).