(SERVER-1922) Set puppet log level from logback
In the early days of Puppetserver, the Puppet log level was maxed out in
order to pass all filtering control to logback. This approach didn't
take into account the potential cost of generating log messages and in
the real world can have performance implications.

This pull request resolves this problem by using the logback
puppetserver log level to set the Puppet log level. Because there is not
a 1:1 mapping of logback levels to Puppet levels the notice log level
cannot be set; if notice level information is needed then the info level
must be set.