(doc) Update CSR gen for intermediate ca doc
The CSR generation script would only print a PEM encoded CSR to stdout
but wouldn't write the CSR to a file; this commit adds logic to write a
file containing the CSR as well.

This commit also fixes up some indentation that would cause some text to
be treated as a code section.