(SERVER-2025) Record metrics for max-queued-requests
This patch adds a new JRuby metric, `queue-limit-hit-meter`, that counts the
number of times the `max-queued-requests` limit was hit and provides
one minute, five minute and fifteen minute measurements on the rate at which
the hits are occurring. This metric can be used to monitor the amount of
traffic that Puppet Server is rejecting due to the `max-queued-request` limit.

The count and five minute rate are also added to the status output for the
JRuby service.