(SERVER-1974) Add node lookup timings to profiler metrics
This patch adds timings captured by `compiler.find_node` to the compiler
metrics returned by the Puppet Profiler. This metric times how long
an External Node Classifier spends looking up a node definition and
can be used to identify when the ENC becomes a bottleneck. Adding this
metric to the profiler's list of compiler metrics makes it available
via the status API for easy consumption.