(SERVER-2022) Update tests to account for PDB
Prior to this a couple of our tests intentionally destroyed our SSL
setup and recreated it. However, with our new installation method
of PDB we are using as it is installed by default. This includes saving
facts to PDB. A partial PKI setup will break fact saving when configured
with PDB.

This updates the tests to account for, and disable while doing unsafe
SSL changes, the configuration that sends facts to PDB.