(SERVER-2022) Don't run pdb tests if pdb isn't there
PDB installation is gated on a list of supported platforms, which
differs(ed?) from the platforms that puppetserver builds and tests for.
Previously the pdb smoke tests were unconditionally run, even if pdb
wouldn't be there. This commit limits them both to only the platforms
listed in the puppetdb_supported_platforms method.