(maint) Strengthen assertions in PuppetDB test
This commit updates the PuppetDB integration test to ensure that the Puppet
Server process is configured to send data to PuppetDB before the test is run.
The assertions in the test are also updated to ensure that facts and a catalog
are submitted in addition to the report and that the timestamps on all
submitted datasets have been updated. This improves upon the old assertion
that just ensured the existence of a report --- which could have been submitted
by a prior test.