(PUP-8141) Set defaults for digest_algorithm, supported_checksum_types
In puppet, in order to better support fips, the digest_algorithm and
supported_checksum_types set defaults based on the value of a fips fact.
This fact does not exist in ruby facter, which puppetserver uses for
unit/integration tests. When the fact doesn't exist, facter is run to
populate all facts. This fails on osx, which requires the cfpropertylist
gem, which won't be present. In order to avoid this problem, this commit
explicitly sets the digest_algorithm and supported_checksum_types if
they aren't already set during startup to avoid requiring the fact at
all.