(SERVER-1954) Roll bidi back to 1.23.1
Starting in bidi 1.25.0, paths that contain invalid URI characters throw
an exception. Unfortunately, puppetserver uses tk-jetty9 and explicitly
tells tk-jetty9 to normalize URIs before handing them to the routing
layer. This means paths with encoded spaces are passed to bidi decoded,
and then have an exception raised, which in turn causes puppetserver to
return a 500. As this is undesireable behavior, this commit rolls the
bidi version in use to 1.23.1, which still allows decoded strings.