(SERVER-2072) Remove logic for specifying 'latest' in test pre-suites
This commit removes the ability to use "latest" in the beaker config to
indicate that the most recent builds of the agent and puppetdb should be
fetched. We now have automation that keeps these versions up to date by
updating SHAs in the beaker config directly when jobs pass, so this
removes brittle hard-coded URLs that went out of date as Jenkins jobs
moved around.