(PE-23484) Update PDB smoke test to work with existing installs
Previously, our PDB installation was using `with_puppet_running_on` which
caused the updated puppet.conf that PDB configured to not be saved
between runs. The PDB smoke test then re-ran the PDB module installation
code to get the puppet.conf updates. That PDB installation code
conflicts with PE, so there were failures when running that smoke test
in our PE pipelines (ie. it was tightly coupled to the FOSS pre-suite).

This updates the FOSS suite to leave the puppet.conf in place once it is
configured for PDB, removes the code from the smoke test that was
attempting to re-apply the module installation code, and updates a test
that was not correctly accounting for PDB being installed and
configured.