(PE-23484) Conditionally disable PDB
Previously, we were always attempting to disable and then re-enable PDB
for our FOSS suite. This causes issue because we do not install and
configure PDB on all our FOSS OSes.

This updates the test in question to only disable and (crucially)
re-enabled PDB when we've actually previously enabled it.