(maint) Revert "(maint) Add build_tar: FALSE to build_defaults"
This reverts commit 755a06aa627bbf2c9f82bc1698ebdbe4f971891c.
The FOSS_ONLY setting (used for shipping) was not properly retrieving tarballs,
but our ship process was still attempting to sign tarballs, causing the ship to
fail. This setting was added to prevent the attempted signing, but now that we
have fixed the underlying problem, this setting is no longer needed.