(SERVER-2024) Clean up puppet install in beaker pre-suite
The logic around installing puppet-agent on the agent nodes during test
setup had suffered a series of refactors that left some nonsense logic
in place. This commit updates that pre-suite to no longer check the
`puppet_version` variable, which a) referred to puppet itself, not the
agent, and b) was unnecessary for install because the version getting
installed is determined by which dev repo gets installed in a previous
step.

The install logic for Windows correctly uses `puppet_build_version` and
was left unchanged.