(maint) Run tests against jruby 9k as well as 1.7
Previously we would only run travis and specs against the default jruby.
This is undesirable as we would like to catch test failures against 9k
before merging, so this commit adds a matrix to the travis configuration
to run against both jruby versions. The Rakefile is also updated to
support running specs against jruby9k.