(maint) Set environment_timeout to 0 in tasks tests
Previously environment_timeout was set to unlimited in the tasks tests.
Due to a puppet bug, we still saw new environments as they were created.
Once that bug was fixed, new environments would not be seen with it set
to unlimited because they would not be created until after puppet had
started up. Now we set it to 0 to ensure new environments are seen. We
also no longer need to generate environment.conf, so that call is
removed from the fixture creation.