(SERVER-2098) Remove jrjackson gem
We ran into some issues with JrJackson's handling Ruby's BigDecimal
type, that causes numbers in scientific notation to be rendered to JSON
as strings, breaking our schema validation. Until this can be resolved,
we can't use the Jackson backend in Puppet Server, so this commit
removes the wrapper gem. MultiJson will now just select the built-in
JSON gem.