(SERVER-2157) Re-bundle JRuby jar
This commit updates puppetserver's build process to start bundling its
JRuby jar again, now that we are only supporting JRuby 9k. If the
JRUBY_JAR setting is used, either at the command line or in the init
config, a warning will be issued indicating that it is no longer needed.