(maint) Only auto-build cljs when building packages
Previously we built our cljs assests every time we started up the repl,
which adds an annoying amount of time to developing on Puppet Server's
main application.

This patch moves to only automatically compiling our cljs assests when
building packages. When doing development on the cljs assests the
developer should be able to manually build the assests with `lein
cljsbuild`.