(SERVER-2192) Use environment name string for listing tasks
When requesting a list of tasks, pass the requested environment as a
string rather than a symbol. The environment cache is primarily used
with string; pass as a string to ensure we re-use a cached environment
if available and avoid registering a cached environment based an a
symbol rather than string name.