(SERVER-1975) Track borrow times per-HTTP request
This commit extends the logic for keeping metrics on JRuby borrow times to
tracking borrow times per-HTTP request. Prior to this commit, only a single
global borrow timer was available which made it difficult to determine what
sorts of requests were bottlenecking a Puppet Server. This commit adds logic
for dynamically creating borrow timers based on the Comidi route ID stored
in the JRuby request reason.

The global timer is kept for backwards compatibility.