(maint) Log errors occurring in JRuby callbacks
This commit refactors the callback function used by the JRubyMetricsService
to include an error trap that logs any exceptions that occur when executing
the callbacks. These exceptions can prevent an instance from being borrowed
or returned to the pool and were previously unlogged. This manifested in
the form of tests that hung indefinitely when test data did not match the
schema expected by functions that compute metrics.