(maint) Normalize HTTP method keywords in tests
This commit fixes up a bunch of mocked Ring requests in the test suite that
were using something other than a lower-case keyword as the request method.
I.e :GET or "get" instead of :get. The Ring library normalizes all incoming
requests to lower-cased keywords and both the Comidi schemas and the Ring
specs expect the :request-method entry in the request map to match a
lower-case keyword.