(maint) Handle incomplete HTTP request maps
This commit updates the JRubyMetricsService to handle incomplete HTTP request
maps when summarizing the reason that a JRuby was borrowed by assigning a
Comidi :route-id of "unknown-http-request" to any Ring request map that lacks
Comidi info. This resolves schema errors in the JRuby metrics handlers arising
from test mocks that do not travel through the Comidi handlers in
the MasterService.