(SERVER-2214) Configure gempath for clojure tests
With the unvendoring of the semantic_puppet gem from Puppet, Puppet now
has a hard dependency on a gem as it is used with the context of
Puppetserver. Previously, we were not installing any gems or configuring
the gempath when running our clojure tests, but with this new
dependency, we need to. This commit pipes the gempath configuration
through to all the places necessary to run unit tests successfully.