(maint) Refactor dev/test gem installation
This commit moves us to using the much faster gem installation that RE
was using during actual packaging. It also moves the
`install-test-gets` script into the dev directory since it's a
development utility and updates `./dev-setup` to call it so there's only
one script call needed to setup a working dev environment.