(SERVER-2231) Publish infrastructure CRL in addition to full CRL
Previously, the CA's CRL would grow with every certificate revoked. In
large environments, especially with ephemeral agents, the size of the
CRL could become problematic.

In practice agents that download the CRL usually only want to have (and
update) a CRL that contains the infrastructure nodes an agent may talk
to.

This patch creates a new setting `enable-infra-crl` which will cause the
CA to look in a file, which lives at "<cadir>/infra_inventory.txt," for
nodes in the Puppet environment that should be considered
"infrastructure" nodes. When an infrastructure node is revoked an
infrastructure node CRL (at "<cadir>/infra_crl.pem") will be updated (in
addition to the usual full CA CRL which is kept on disk for auditing
purposes). When agents check in to download a CRL the CA will serve the
infra CRL.

Also includes updates to some mock CA structures used during unit
tests to account for new config settings