(SERVER-2278) Add allow-subject-alt-names setting to puppetserver.conf Before this commit we completely disallowed signing certificate requests with subject alternative names from Puppetserver's certificate_status endpoint.  Now users can choose to allow signing of certs with subject alt names by setting allow-subject-alt-names to true in the puppetserver .conf file.  Setting is false by default.
Note that in Puppet Server 5, this setting does NOT replace the
`allow_dns_altnames` flag of the `puppet cert sign` command. When using
this command, it is still necessary to use the flag to sign certs with
alt names.

This is a cherry-pick of eb062d18844aece82b2fa22bef40d2e381ac1e8d.