(SERVER-2290) Add a setting to enable signing CSRs with auth extensions Before this commit we completely disallowed signing certificate requests with authorization extensions from Puppetserver's certificate_status endpoint.  Now users can choose to allow signing of certs with auth extensions by setting allow-authorization-extensions to true in the puppetserver .conf file.  Setting is false by default.
Note that in Puppet 5, this setting does not replace the
`allow-authorization-extenstions` flag on `puppet cert sign`. It is
still necessary to pass the flag when using that command, even if this
setting is enabled for Puppet Server.

This is a cherry-pick of 1a9345bb340ec30ca535e569c7f8283a611a3a2b.