(SERVER-2287) Add OID for custom CLI auth extension
This commit adds a custom extension in Puppet's arc that is checked when
allowing access to the certificate_status(es) endpoint. This extension
should only be added to the master's cert by the CLI tool's generate
command, and it is disallowed when signing other certs with this CA.

This is a cherry-pick of d7d1d79862ade7acacb75e3ea5400f948da18abb from
the 6.0.x branch.