(SERVER-2302) Allow access to all status endpoints
Previously, we added the ability for the status service to honor tk-auth
configs. To provide backwards compatability we opened up the default
tk-auth rules for the status service to be un-authenticated. However, in
that work we only opened up /status/v1/services, to allow calls to all
endpoints mounted under the status service this relaxes the path to only
be /status/v1.

During the previous work we also updated the tk-status service
documentation, however we missed documenting this change in our public
Puppet Server documentation. This includes that update as well.