(SERVER-2218) Use valid class names in compilation tests
Previously we would create dummy classes in throwaway environments that
did not follow best practices for class naming. These became
deprecations at the end of the Puppet 5 series and will be errors in
Puppet 6.

This updates our mock class generation and expected output helpers to
create and test for class names that follow Puppet's best practices (and
won't err in the immediate future).