(maint) Remove extra unwrapping of SSL exceptin in test
We started unwraping Ruby exceptions (and updated this assertion) in
968d39b (which changed the stable branch). 66e526e also changed these
tests in master branch. It looks like in a merge up the expectations
from master were taken with the utils of stable. This does cause
problems because current Java raises several wrapped SSL ex., however
in jdk11 only one SSL exception is raised.