(maint) remove cljs code and assets
The developer dashboard receives little use and the continued
maintenance of it has become enough of a burden that we've re-evaluated
how much value we and the community are receiving from it. That
evaluation pointed out that our Support and Services departments own a
solution that leverages graphite within a docker container to query
a larger collection of data from the status endpoint. They also own a
polling job that can be installed to ensure historical data. They have
open sourced these tools and they are now the recommended way for users
to solve the problems that this dashboard solves.