(maint) Specify `server` setting puppet.conf
Previously, we did not specify the server setting in the puppet.conf but
used `--server <server>` in any invocation that required it. The
`puppetserver ca` cli does not have a `--server` flag like puppet does
(it expects you to have your agent configured to actually talk to your
master correctly). Consequently we need to have the server value set in
our puppet.conf (the default of "puppet" is used by our internal Ops
infra). Previously, the puppetserver ca tool used a default server value
of `certname` rather than `puppet` so it didn't require this.