(maint) Update metrics test mocks for new scheduler
The tk-scheduler returns an opaque value when scheduling a job, that
value can then be used to stop said job.

In our metrics tests we mock out scheduling so we can control the
sampling of metrics within our tests. The garbage values that our mock
creates are saved by the service and then used to try and stop any
scheduled job on test shutdown.

Invalid values before were silently ignored, however now the scheduler
throws an error when attempting to stop an invalid job. But since this
occurs during test teardown it does not cause a test error.

This patch mocks out the scheduler stop-job protocol as well so it will
ignore the garbage ids our mock scheduling creates.