(maint) Remove leftover lein dependency
As part of removing the developer dashboard, we removed the need to
depend on lein as a dev dependency. It was removed from project.clj at
that time, but not from dev_tools.clj, meaning that the repl can no
longer be started on the 5.3.x series. In the 6.x series, this dep was
removed as part of the move to figwheel-sidecar. This commit removes it
from dev_tools.clj in the older branch.