Add new catalog endpoint (#1962)
This commit adds the clojure skeleton for a new catalog endpoint, designed to be more flexible by taking many params to control where inputs to compilation come from and whether or not to store facts and catalog. This can be used to make requests for arbitrary nodes, by any requester (not just the node itself).