(SERVER-2444) Avoid global state when using terminus method
This instantiates the PDB Facts terminus directly and invokes its `find`
method with a manually-created Indirector::Request object. This does not
rely on or interfere with the global state in Indirector::Indirection.

Note that by avoiding the logic in Indirector::Indirection, we lose
automatic caching that would normally be configured by the indirector.

Also note that terminii are usually loaded and created by the
Autoloader. I'm not completely clear if there is something else we're
losing (safety around missing terminii?) by skipping that part.