(SERVER-2457) Collect Indirector information
Previously, when compiling a v4 catalog, we directly
created a PDB termini for retrieving facts after checking on
Puppet.settings. This was to avoid the Indirector during each compile
since it is both confusing and not thread safe.

However, the Indirector may be configured outside of the puppet.conf
by using the routes.yaml file (which PuppetDB actually recommends).
This patch updates us to using the Indirector so we do not have to
re-write its routes.yaml processing and Puppet.settings merging.

However, to avoid using it for each request (which we would love to make
thread safe at some point in the future) we intended to inspect the
indirector only once, when Puppet/JRuby is first initialized. We can
then reference that information in a thread safe manner for each request.

Furthermore, by isolating implementation of inspecting the Indirector we
hope to easily move to using whatever may replace it in the future
(cross our fingers that future comes).