(SERVER-115) Avoid potential corruption by serializing crl writes
Note that this is just a trivial mitigation -- it does not prevent
corruption that might result from a server crash while writing, nor
does it prevent simultaneous crl read operations from seeing partial
updates.

To fix those, we might rework the code to make the updates to a temp
file while holding the lock, and then use Files/move to ATOMIC_MOVE
the update into place.