(SERVER-2475) Return logs from compilation with response
This commit adds code to wrap the catalog compilation that captures any
logging emitted during the compilation and returns it along with the
catalog. Normally, this logging would just go into the main puppetserver
log, and would be interleaved with any other compilations going on,
making it very hard to tell where the messages were coming from. This
way, the logs are tied to the catalog whose compilation produced them.

Note that this means that the logs will NOT also go to Puppet Server's
normal log destination.